move: 카드 파일 구조
