+++ src/handler/card/debuffs/containmentUnitCard.handler.js
@@ -1,4 +1,4 @@
-import { Packets } from '../../init/loadProtos.js';
+import { Packets } from '../../../init/loadProtos.js';
 
 export const containmentUnitCardHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
   targetUser.characterData.debuffs.push(Packets.CardType.CONTAINMENT_UNIT);
