+++ src/handler/card/debuffs/satelliteTargetCard.handler.js
@@ -1,4 +1,4 @@
-import { Packets } from '../../init/loadProtos.js';
+import { Packets } from '../../../init/loadProtos.js';
 
 export const satelliteTargetCardHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
   targetUser.characterData.debuffs.push(Packets.CardType.SATELLITE_TARGET);
