+++ src/handler/card/equips/autoShieldCard.handler.js
@@ -1,4 +1,4 @@
-import equipCardNotification from '../../utils/notification/equipCard.notification.js';
+import equipCardNotification from '../../../utils/notification/equipCard.notification.js';
 
 export const autoShieldHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
   cardUsingUser.addEquip(useCardType);
