+++ src/handler/card/equips/radarCard.handler.js
@@ -1,4 +1,4 @@
-import equipCardNotification from '../../utils/notification/equipCard.notification.js';
+import equipCardNotification from '../../../utils/notification/equipCard.notification.js';
 
 export const radarHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
   cardUsingUser.addEquip(useCardType);
