+++ src/handler/card/normal/absorbCard.handler.js
@@ -1,10 +1,7 @@
 import {
-    getStateNormal,
     getStateAbsorbing,
     getStateAbsorbTarget,
-  } from '../../constants/stateType.js';
-  import { PACKET_TYPE } from '../../constants/header.js';
-  import { createResponse } from '../../utils/response/createResponse.js';
+  } from '../../../constants/stateType.js';
 
 export const absorbCardHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
     cardUsingUser.setCharacterState(getStateAbsorbing(targetUser.id));
