+++ src/handler/card/normal/deathMatchCard.handler.js
@@ -2,8 +2,8 @@ import {
   getStateNormal,
   getStateDeathInitShooter,
   getStateDeathInitTarget,
-} from '../../constants/stateType.js';
-import userUpdateNotification from '../../utils/notification/userUpdate.notification.js';
+} from '../../../constants/stateType.js';
+import userUpdateNotification from '../../../utils/notification/userUpdate.notification.js';
 
 // 플레이어 한명을 지정하여 번갈아가며 빵야!카드를 낸다. 빵야!를 못내면 체력 1 소모  타겟 : 목록에서 선택  방어 카드 : 빵야!
 export const deathMatchCardHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
