+++ src/handler/card/normal/hallucinationCard.handler.js
@@ -1,8 +1,7 @@
 import {
-    getStateNormal,
     getStateHallucinating,
     getStateHallucinationTarget,    
-  } from '../../constants/stateType.js';
+  } from '../../../constants/stateType.js';
 
 export const hallucinationCardHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
     cardUsingUser.setCharacterState(getStateHallucinating(targetUser.id));
