+++ src/handler/card/weapons/sniperGunCard.handler.js
@@ -1,4 +1,4 @@
-import { Packets } from '../../init/loadProtos.js';
+import { Packets } from '../../../init/loadProtos.js';
 
 export const sniperGunCardHandler = (cardUsingUser, targetUser, currentGame, useCardType) => {
   if (cardUsingUser.characterData.weapon !== 0) cardUsingUser.unequipWepon();
